Extract common components of mobile & desktop views
